git: Don't trust ctime, just in case some stupid program is accessing the same files
Git relies on ctimes by default to check if files changed. If in the middle of
an interactive rebase someone/something touches the files, git gets really
confused, and if we aren't careful, commits can be lost by accident.

Did I tell you I'm not a fan of background indexing daemons on my machines?